requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.6.4
python-dotenv==1.0.1
alembic==1.13.1
websockets==12.0
aiofiles==23.2.1
Pillow==10.2.0
